Fix NoMethodError in dev.rb when skip-active-job
config.active_jobs.verbose_enqueue_logs was [added][1] to the
development.rb template, but it was put inside the skip-active-record
block. This means that it will be included in generated development.rb
even when Active Job is skipped.

This commit fixes this issue by moving it outside of the
skip-active-record check and into its own skip-active-job check.

[1]: f6d56fed27cfd11aa7161129e2c59a972f90b76b