Eager evaluate relation in composite case
Normally, it's valid syntax to pass the predicate builder a mapping from
primary key to a record, to an id, or to a relation.

With the composite case, there's a limitation on the types of syntax
supported. Namely, the only case we support right now is mapping a tuple
of columns to a tuple of corresponding values. For now, it's sufficient
to extract the ids instead of evaluating the SQL at a later time.